requirements.txt
playwright
requests
beautifulsoup4
cssutils
Pillow
tqdm
scikit-learn
numpy
python-dotenv